[Feat #152] 조건부 렌더링을 통한 로딩창 구현
