+++ components/mafia/Loading.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Loading = () => {
+  return (
+    <>
+      <div style={{ backgroundColor: "red", width: "100vw", height: "100vh" }}>나가는 중</div>
+      <div className="via-fuchsia-500, w-4/5, h-3/4">dfssdfsafasdfsdfsddfs</div>
+    </>
+  );
+};
+
+export default Loading;
